+++ backend/src/api/internal/services/gift.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Iterable, List, Optional
 
 from django.db import IntegrityError, transaction
@@ -11,11 +12,11 @@ def get_gifts(source: Profile) -> List[Transaction]:
     return Transaction.objects.filter(source=source).all()
 
 
-def is_enough_money(user: Profile, accrual: int) -> bool:
+def is_enough_money(user: Profile, accrual: Decimal) -> bool:
     return accrual <= user.balance
 
 
-def try_transfer(source: Profile, destination: Profile, description: str, accrual: int) -> Optional[Transaction]:
+def try_transfer(source: Profile, destination: Profile, description: str, accrual: Decimal) -> Optional[Transaction]:
     if not is_enough_money(source, accrual):
         return None
 
@@ -35,7 +36,7 @@ def try_transfer(source: Profile, destination: Profile, description: str, accrua
         return None
 
 
-def try_gift(source: Profile, destinations: Iterable[Profile], description: str, accrual: int) -> List[Transaction]:
+def try_gift(source: Profile, destinations: Iterable[Profile], description: str, accrual: Decimal) -> List[Transaction]:
     transactions = []
     for destination in destinations:
         transaction_ = try_transfer(source, destination, description, accrual)
@@ -45,7 +46,7 @@ def try_gift(source: Profile, destinations: Iterable[Profile], description: str,
     return transactions
 
 
-def _get_transaction(source: Profile, destination: Profile, description: str, accrual: int) -> Transaction:
+def _get_transaction(source: Profile, destination: Profile, description: str, accrual: Decimal) -> Transaction:
     return Transaction.objects.create(
         type=TransactionTypes.COIN_GIFTING,
         source=source,
